plugin: extract Plugin subsystem specific fields into a separate PluginContext struct
There's an ugly level of coupling between GlobalConfig and the handling
of Plugins, basically plugin.c is just a set of methods related to
GlobalConfig, as most functions take a GlobalConfig *cfg as an argument.

This is a start of a series that aims at removing this coupling and
introducing a GlobalConfig -> Plugin dependency in a single direction.

This time it's only a first step to introduce the PluginContext struct,
for now in cfg.h that will later be moved to plugin.h as the interface
between the two classes get clearer.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>