cfg: publish initial module  directory
module-path is actually a list of directories to look for modules, yet
it was used by the java based plugins simply as a base path.

To fix this situation we introduce a `module-install-dir` define, that
contains the installation path for modules, which is the
default value for module-path as well. Still, module-install-dir will
not be changed, thus it is safe to use that to formulate actual paths
on the filesystem.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>