cfg-lexer: move "globals" hash table to GlobalConfig
It can be argued that the set of @defined variables actually belong
to the configuration, and there are a number of cases where GlobalConfig
was querying, changing elements.

Also, CfgBlock needs to access to the "globals" hash and the fact that
it was stored in CfgLexer was skewing the interface.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>